+++ biokit/__init__.py
@@ -9,13 +9,19 @@ except:
 from easydev import CustomConfig
 biokitPATH = CustomConfig("biokit").user_config_dir
 
+import colorlog as logger
+def biokit_debug_level(level="WARNING"):
+    """A deubg level setter at top level of the library"""
+    assert level in ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]
+    logging_level = getattr(logger.logging.logging, level)
+    logger.getLogger().setLevel(logging_level)
+    
 
 from biokit import viz
-from biokit import io
-from biokit import services
-
 from biokit.viz import *
-from biokit.services import EUtils, KEGG, UniProt
+
+from biokit import io
+
 
 from biokit import stats
 from biokit.network import *
@@ -29,14 +35,15 @@ from biokit.goid import *
 from biokit import taxonomy
 from biokit.taxonomy import Taxonomy
 
+from biokit import converters
 
 
 def biokit_data(filename, where=None):
     """Simple utilities to retrieve data sets from biokit/data directory"""
     import os
     import easydev
-    sequana_path = easydev.get_package_location('biokit')
-    share = os.sep.join([sequana_path , "biokit", 'data'])
+    biokit_path = easydev.get_package_location('biokit')
+    share = os.sep.join([biokit_path , "biokit", 'data'])
     # in the code one may use / or \ 
     if where:
         filename = os.sep.join([share, where, filename])