+++ biokit/rtools/r4python.py
@@ -22,11 +22,11 @@ r4python = r"""
 .getRvalue4Python__ <- function(x, use_dict=NULL) {
 
     headstr <- 'numpy.array('
-    tailstr <- ')' 
+    tailstr <- ')'
 
     SpecialLocs <- function(x) { # find locations of special values: NULL, NA, NaN, Inf
         rlt <- list()
-        
+
             idx <- which(is.null(x) | is.na(x))
             if (length(idx) > 0) rlt$None <- idx
             idx <- which(is.nan(x))
@@ -39,7 +39,7 @@ r4python = r"""
                 iidx <- which(v < 0)
                 if (length(iidx) > 0) rlt['-numpy.Inf'] <- idx[iidx]
                 }
-            
+
         return(rlt)
         }
 
@@ -109,6 +109,10 @@ r4python = r"""
                 special_locs <- SpecialLocs(xi)
                 if (is.character(xi)) {
                     ctp[i] <- sprintf('("%s", "|S%d")', cnms[i], if (length(xi) > 0) max(nchar(xi)) else 0 )
+                    # Hack TC July 2017 S%d above was sometime replaced by SNA
+                    # in Python3. Let us use object for now
+                    ctp[i] <- sprintf('("%s", "object")', cnms[i], 50 )
+
                     xi <- paste('"', xi, '"', sep='') }
                 else if (is.logical(xi)) {
                     xi <- ifelse(xi, 'True', 'False')
@@ -124,7 +128,7 @@ r4python = r"""
                     ctp[i] <- paste('("', cnms[i], '", "<G")') }
                 if (length(special_locs) > 0) xi <- SpecialVals(xi, special_locs)
                 if (nrow(x) > 0) x[[i]] <- xi }
-            tailstr <- paste(', dtype=[', paste(ctp, collapse=','), ']', tailstr, sep='') 
+            tailstr <- paste(', dtype=[', paste(ctp, collapse=','), ']', tailstr, sep='')
 
         x <- as.matrix(x)
         x <- apply(x, 1, function(r) paste('(', paste(r, collapse=','), if(length(r)<2) ',)' else ')', sep=''))
@@ -190,5 +194,5 @@ r4python = r"""
 
     if(identical(.Platform$OS.type, 'windows')) .addLibs()
     rm(.addLibs)
-    
+
 """