+++ test/converters/test_utils.py
@@ -1,0 +1,9 @@
+from biokit.converters import utils
+
+
+def test_utils():
+
+
+    res = utils.MapperRegistry()
+    assert len(res)
+    assert "bam2bed" in res